refactor: move the type, constant, and model to their locations
