+++ apps/web/src/services/alert-manager/components/AlertMainAcknowledgeModal.vue
@@ -29,6 +29,7 @@ import { PButtonModal, PCheckbox } from '@spaceone/design-system';
 
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 
+import { ALERT_STATE } from '@/schema/monitoring/alert/constants';
 import { store } from '@/store';
 import { i18n } from '@/translations';
 
@@ -37,8 +38,8 @@ import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
 import ErrorHandler from '@/common/composables/error/errorHandler';
 import { useProxyValue } from '@/common/composables/proxy-state';
 
-import { ALERT_STATE } from '@/services/alert-manager/constants/alert-constant';
 import type { AlertStateUpdateParams } from '@/services/alert-manager/types/alert-type';
+
 
 export default {
     name: 'AlertMainAcknowledgeModal',
