+++ apps/web/src/services/alert-manager/composables/alert-urgency-i18n.ts
@@ -2,9 +2,9 @@ import type { ComputedRef } from 'vue';
 import { computed } from 'vue';
 import type { TranslateResult } from 'vue-i18n';
 
+import type { AlertUrgency } from '@/schema/monitoring/alert/model';
 import { i18n } from '@/translations';
 
-import type { AlertUrgency } from '@/services/alert-manager/constants/alert-constant';
 
 export const useAlertUrgencyI18n = (): ComputedRef<Record<AlertUrgency, TranslateResult>> => computed(() => ({
     ALL: i18n.t('MONITORING.ALERT.ALERT_LIST.ALL'),
